+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes for Meilisearch Connect
 
+# Unreleased
+
+- Improved utility view when no indexes or only search-only indexes are present
+
 # 2.0.0 - 2026-03-18
 
 - Add indexed document tracking